[native_assets_cli] Move non-`build.dart` code to `package:native_assets_builder`
