+++ pkgs/native_assets_builder/pubspec.yaml
@@ -12,9 +12,10 @@ dependencies:
   logging: ^1.2.0
   native_assets_cli: ^0.3.2
   package_config: ^2.1.0
+  yaml: ^3.1.2
+  yaml_edit: ^2.1.0
 
 dev_dependencies:
   dart_flutter_team_lints: ^2.1.1
   file_testing: ^3.0.0
   test: ^1.24.3
-  yaml: ^3.1.2
