+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,7 +1,12 @@
-## 0.3.3-wip
+## 0.4.0-wip
 
 - Added [example/use_dart_api/](example/use_dart_api/) detailing how to use
   `dart_api_dl.h` from the Dart SDK in native code.
+- **Breaking change** Moved code not used in `build.dart` to
+  `package:native_assets_builder`.
+  `AssetRelativePath` is only defined inside `toNativeAssetsFile()` which is
+  passed to the VM, not inside `build.dart` output.
+
 
 ## 0.3.2
